# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FormalAP1",
        "version": "1.0.0"
    },
    "paths": {
        "/api/circuits": {
            "get": {
                "tags": [
                    "Circuits"
                ],
                "description": "Return all circuit with their informations.",
                "operationId": "getAllCircuit",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/circuits/{circuit}": {
            "get": {
                "tags": [
                    "Circuits"
                ],
                "description": "Return circuit by ID.",
                "operationId": "getAllCircuit",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Circuits"
                ],
                "description": "Update circuit by Id.",
                "operationId": "UpdateCircuit",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Circuits"
                ],
                "description": "Update circuit by Id.",
                "operationId": "UpdateCircuit",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/circuits/": {
            "post": {
                "tags": [
                    "Circuits"
                ],
                "description": "Create new circuit.",
                "operationId": "CreateCircuit",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/Constructors": {
            "get": {
                "tags": [
                    "Constructors"
                ],
                "description": "Return all Constructor with their informations.",
                "operationId": "getAllConstructor",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/constructors/{constructor}": {
            "get": {
                "tags": [
                    "Constructors"
                ],
                "description": "Return constructor by ID.",
                "operationId": "getAllConstructor",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Constructors"
                ],
                "description": "Update constructor by Id.",
                "operationId": "UpdateConstructor",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Constructors"
                ],
                "description": "Update constructor by Id.",
                "operationId": "UpdateConstructor",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/constructors/": {
            "post": {
                "tags": [
                    "Constructors"
                ],
                "description": "Create new constructor.",
                "operationId": "CreateConstructor",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
...
    },
    "security": [
        {
            "constants": {
                "L5_SWAGGER_CONST_HOST": "http://127.0.0.1:8000/api"
            }
        }
    ]
}